Upgraded Llithuanian vocabulary. Few optimizations to bruteforce.
Lithuanian vocabulary updated to 83k words. Bruteforce was optimized in a way to only read words from dictionary that match the length of words in encrypted text, in that way we dont' have to compare all the words in dictionary.  Added dictionary file name printing to console. Added color header, to make colored prints in the console.